chore. Restructure directory files
