+++ .github/README_ES.md
@@ -1,6 +1,6 @@
 # ![logo](https://raw.githubusercontent.com/azerothcore/azerothcore.github.io/master/images/logo-github.png) AzerothCore
 
-## # mod-aoe-loot
+# mod-aoe-loot
 
 [English](README.md) | [Español](README_ES.md)
 
@@ -8,47 +8,14 @@
 
 [![Build Status](https://github.com/azerothcore/mod-aoe-loot/workflows/core-build/badge.svg?branch=master&event=push)](https://github.com/azerothcore/mod-aoe-loot)
 
-![Example](example.gif)
+# Nota importante
 
-## Notas importantes
+El modulo, se encuentra actualmente en desarrollo. Si bien, creemos que funciona en casi todos los ámbitos, dado que hemos realizado pruebas y también recibido reportes, de personas que jugaban en grupo, personas en solitario, personas dentro de una mazmorra (ya sea que ingresaron caminando o utilizaron el modulo de `mod-solo-lfg`), no descartamos que pueda llegar a surgir algún error, por lo que es importante, que nos ayudes a probar el contenido, y creando un reporte dentro de los issue del repositorio, en caso de que encuentres algún posible fallo.
 
-El modulo, actualmente se encuentra en desarrollo, pero es funcional. Deberia de ser compatible al menos, con el modulo: mod-solo-lfg. Se puede realizar el loot en area, tanto en grupo, como en solitario, y funciona tanto al obtener un item, como al obtener el oro, por lo que la opcion de autodestoje del cliente, podria estar habilitada. En el fichero de configuracion, existen algunas variables, que pueden ser modificadas, como el hecho de enviar el item por correo en caso de tener las bolsas llenas, o aumentar el rango de busqueda de cuerpos. Al picar una mena, puede que tambien obtengas todos los items de los cuerpos, con cofres he realizado la prueba y no funciona, pero con menas, tengo entendido que si, faltaria ver si tambien admite la profesion de herboristeria en caso de tener algun material cerca de los cuerpos.
+Para funcionar, el modulo requiere de versiones actualizadas del emulador, dado que se utilizan algunos hooks que existen desde hace tiempo, pero también se tuvo que crear unos hooks para el mismo. Por lo que corrobore su versión del emulador, con el comando `.server info` y asegúrese de utilizar la versión mas actualizada del mismo. Dentro de la carpeta `data/sql` se encuentra un simple script que añade traducciones del mensaje de bienvenida. Puede añadir los idiomas restantes, y el script, debería de ejecutarse automáticamente, si tiene habilitadas las opciones por defecto dentro del `worldserver.conf`, caso contrario, puede que tenga que ejecutarlas de forma manual.
 
-El tiempo, que un cuerpo permanece en el mundo, luego de no tener items, se determina en el fichero de configuracion del worldserver.conf, por lo que al hacer loot en area, es necesario quizas, aumentar un poco ese tiempo, para darle tiempo a las personas que tenga la habilidad de desuello, el tiempo suficiente, para poder obtener los materiales de dicha habilidad, sin que el cuerpo se borre demasiado rapido.
+# ¿Como se utiliza?
 
-```
-#
-#    Corpse.Decay.NORMAL
-#    Corpse.Decay.RARE
-#    Corpse.Decay.ELITE
-#    Corpse.Decay.RAREELITE
-#    Corpse.Decay.WORLDBOSS
-#        Description: Time (in seconds) until creature corpse will decay if not looted or skinned.
-#        Default:     60   - (1 Minute, Corpse.Decay.NORMAL)
-#                     300  - (5 Minutes, Corpse.Decay.RARE)
-#                     300  - (5 Minutes, Corpse.Decay.ELITE)
-#                     300  - (5 Minutes, Corpse.Decay.RAREELITE)
-#                     3600 - (1 Hour, Corpse.Decay.WORLDBOSS)
+Clona el repositorio dentro de la carpeta modules del emulador, o descarga el zip y descomprimelo. Ten en cuenta, que si utilizas el .zip, es probable, que te añada al final de la carpeta descomprimida, un `-` y a continuación, el nombre de la rama que hayas descargado. Te recomendamos, clonar el repositorio, para poder, en caso de existir, descargar actualizaciones fácilmente sin la necesidad de tener que descargar todo el contenido nuevamente.
 
-Corpse.Decay.NORMAL    = 60
-Corpse.Decay.RARE      = 300
-Corpse.Decay.ELITE     = 300
-Corpse.Decay.RAREELITE = 300
-Corpse.Decay.WORLDBOSS = 3600
-
-#
-#    Rate.Corpse.Decay.Looted
-#        Description: Multiplier for Corpse.Decay.* to configure how long creature corpses stay
-#                     after they have been looted.
-#         Default:    0.5
-
-Rate.Corpse.Decay.Looted = 0.5
-```
-
-Este módulo requiere el siguiente pull request
-
-https://github.com/azerothcore/azerothcore-wotlk/pull/16589
-
-## Descripción
-
-El objetivo de este módulo es permitir a los jugadores obtener todos los objetos de varios npc's dentro de un cierto rango de distancia.
+Gracias a todas las personas, que de alguna forma u otra participaron y participan activamente de este modulo, para que el mismo, pueda continuar actualizado y funcionando. Si eres colaborador o te interesa serlo, no dudes en abrir un pull request, con el contenido que te gustaría que tuviera el modulo o para arreglar algún posible fallo.
